+++ samples/EasyOBJDCalibrateSample.java
@@ -13,10 +13,10 @@ import android.annotation.SuppressLint;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.easyobd.ClusterInfo.Ball;
-import org.firstinspires.ftc.easyobd.EasyOBJD;
-import org.firstinspires.ftc.easyobd.EasyOBJDCalibration;
-import org.firstinspires.ftc.easyobd.EasyOBJDPipeline;
+import org.firstinspires.ftc.easyobjd.ClusterInfo.Ball;
+import org.firstinspires.ftc.easyobjd.EasyOBJD;
+import org.firstinspires.ftc.easyobjd.EasyOBJDCalibration;
+import org.firstinspires.ftc.easyobjd.EasyOBJDPipeline;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;