+++ samples/EasyOBJDTuner.java
@@ -13,9 +13,9 @@ import android.annotation.SuppressLint;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.easyobd.EasyOBJD;
-import org.firstinspires.ftc.easyobd.EasyOBJDPipeline;
-import org.firstinspires.ftc.easyobd.EasyOBJD.OverlayMode;
+import org.firstinspires.ftc.easyobjd.EasyOBJD;
+import org.firstinspires.ftc.easyobjd.EasyOBJDPipeline;
+import org.firstinspires.ftc.easyobjd.OverlayMode;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;