+++ src/test/java/org/firstinspires/ftc/easyobjd/GridClusteringTest.java
@@ -6,7 +6,7 @@
  * EasyOBJD - FTC EasyOpenCV object detection
  * https://github.com/IamAki123/EasyOBJD
  */
-package org.firstinspires.ftc.easyobd;
+package org.firstinspires.ftc.easyobjd;
 
 import org.junit.Test;