+++ src/test/java/org/firstinspires/ftc/easyobjd/KnownGeometries.java
@@ -6,7 +6,7 @@
  * EasyOBJD - FTC EasyOpenCV object detection
  * https://github.com/IamAki123/EasyOBJD
  */
-package org.firstinspires.ftc.easyobd;
+package org.firstinspires.ftc.easyobjd;
 
 /**
  * Sample / docs numbers. Keep these aligned with